build/drm: Allow building from scratch by not using pkgconfig for internal libs
Using pkgconfig for internal libs turns out to be not a really good idea. It
works fine if you already have an efl install with the needed ecore-drm.pc
file but it will fail if you build from scratch.

We already have a m4 macro for these internal dependencies. Make use of it
for the evas drm engine depending on ecore-drm.

Fixes T1432