Configure.ac: Free up the "efl" name for usage.
At the moment we use the fake "efl" library as a dependency for
everything and use it as a way to have global cflags and lib deps. This
is bad as we'd like to have a "real" libefl.

I changed EFL to EFLALL as the new name. Easy to change to something
else if anyone finds this name lacking. It's internal.