Revert "configure.ac: generate Makevars in the correct directory even with out-of-tree build"
This reverts commit dd37d2bc07f839cd3c74b02e713ffc6534af71b2.

This breaks make distcheck. Looking at this commit I really wonder if it does
anything good. It seems to work for po_makefile_in. It also breaks for
distcheck which is using out of tree builds in the _build folder.

If someone can explain me why something like this is needed for makevars I
want to hear it. getting a fix in that does not break distcheck would be fine
I guess.