build: Avoid -shared LDFLAG to allow me to have a working distcheck
The -shared option in LDFLAG somehow is poisoning the configure run inside
distcheck. This results in the compiler test getting a segfault and failing.

I would really like to understand why this is needed to be set here manually
anyway. Same for fPIC and dPIC. I would expect that we leave this to the user
compiling the code or the packager. Skimming through the log only brought up
fixing c++11 globally. Not really a commit message that helps me to
understand why this is needed...