shot.sh: Set ELM_DISPLAY to bring the elm shot engine back to life
As a small fallout from e7dbe063c9c044f072bacafd31ab872f6210788b we now
need to set ELM_DISPLAY with our system has something like $DISPLAY or
$WAYLAND_DISPLAY set. Elm uses that now to find out which window system
to use.

In the case of daing make doc we want it forced to shot even if we run
under X. With ELM_DISPLAY we do that. Thanks goes to raster for the
explanantion and git bisect for pinning it down.