elm win - fix invalid use of elm engine config per window
lots of code unsing ENGINE_COMPARE where it should get the engine from
the ecore_evas directly - it happened to work as we happened to have
apps that have windows only for x11, or wayland etc, but not both in
an app. this was wrong, so remove the engine compare macro and do it
properly. no laziness.