elm - do not do x11 main/config stuff without DISPLAY set
the engine checks are also broken - but that's a much bigger change,
so this fixes elm config and main code to not do their x11 work if no
DISPLAY env var is set. this is right for a wayland world where you
don't want DISPLAY but WAYLAND_DISPLAY set and used. this fixes
several ERRs that should never have hapapned in this setup.