ecore-imf - fix broken state on non-x11 environments
this fixes ecore-imf to auto-pick scim or wayland vased in DISPLAY and
WALAND_DISPLAY env vars. it's not a perfect/good/correct fix. it
should be picking based on canvas engine type, but that is a far more
invasive effort, so for release, do this simple one that brings back
wayland input panel and keeps x11 input methods going as long as app
is either x11 (DISPLAY) or wayland (WAYLAND_DISPLAY). if both are set,
it'll get it wrong. as i said - that's a more invasive fix.