autotools: proper fix for generated Makevars and distcheck
Reverts 21da4a54545cd6871d1d888566ce9fe488afef67.

It is needed to generate Makevars in-tree even when building out-of-tree because of
how Autotools work. However, distcheck doesn't properly remove the Makevars file in
the generated distdir and makes po/ read only, preventing the build system from
generating an up-to-date version of Makevars. This commit adds the required hooks
needed to fix this behavior.