xpm - colors cant be 96 bytes long.
since colors are hex, 96 bytes of hex would be 32bytes per r, g and b,
and that would be 128 bits per  r g and b and that just is never seen
or used. also coverity is right in CID 1193201 that we just can't
shift that many, so go down to max 16 bits per rgb which is sensible
and able to be done. (thats 12 bytes max for hex of color).