Revert "efl: Fixed unwanted configure.ac warning."
This reverts commit bf8aba5f9f52dd2ce8d0c292b54bae0c44d99462.

This warning is actually wanted. We do want to know when things get deprecated and
not discover that to late. This warning come from the use of gettext 0.17. Once we
move out of it, we will be fine. As a reminder and for keeping track of other
future deprecated macro, we should never use that flag !

Note: This is the second time I revert such a patch, I would really like people
stop disabling warning with this nasty work around.