elm engine choice and logic rationalize
now that we have ELM_DISPLAY and ELM_ACCEL we have little need for a
lot of the elm engine stuff and frankly a lot was broken if you used
multiple engines etc. so this fixes a lot of it nd removes a lot of
cruft. i think i got it all right, but i may have oppsied. this also
remo9ves engine support for engines long gone from evas.