Object Items: Adapt del callback to Eo.
Because of widget item design, it is hard to make it fit to Eo as it
should be:
- _eo_del_cb should have the callback pointer as data but pre_notify
function forces its storage into the item data because we need to be
able to call the cb before del is really triggered.
- only one cb is supported for an item, meaning replacing it would mean
we have to save the function pointer inside the item, delete the
callback with the previous function as data and then add the new
callback. Instead of it, we just add our own wrapper as Eo callback that
will invoke the function pointer stored in the item data.