eeze: Fix issue with eeze_udev_find_by_type not finding devices when name is NULL.
Summary: This was totally wrong !! eeze_udev_find_by_type can be
passed EEZE_UDEV_TYPE_NONE to list all devices and it can also be
passed name == NULL to find All devices listed of a certain type. To
check for !etype && !name is just plain Wrong !! We have checks below
for a valid name (and then do a comparison).

@fix

Signed-off-by: Chris Michael <cp.michael@samsung.com>