autotools: modified autogen.sh to be able to build from any directory.
Summary:
Now we cannot have a designated build directory for building EFL. We have to build from the same source folder only. So have modified autogen.sh script
so that we can build from any directory. This is similar to the one which I had submitted for Enlightenment and Elementary.

Signed-off-by: Srivardhan Hebbar <sri.hebbar@samsung.com>

Reviewers: devilhorns, cedric, raster

Subscribers: cedric

Differential Revision: https://phab.enlightenment.org/D1480

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>