evas - render2 renders conecnt for the first time... wow...
this is an unused code path UNLESS you:

export EVAS_RENDER2=1

so this will affect nothing unless you enable this and test it. it's
designe dto be done this way so we can slowly bring it in and bring it
up to speed and peolpe can easily test it and so on.

so now - content is rendered. most content is actually. only thing
that breaks that i can see so far is mapped smart objects.