eina: remove @addtogroup in eina_inline_mempool.x: the .h already has it.
It looks like doxygen doesn't appreciate having a group inside a group of
the same name and simply ignores any documentation coming from the parent
one.

Since the .x is included from the .h, it is always already inside the
group so simply remove the directive from the .x.

This fixes the documentation for all the functions which are declared and
documented in the .h and implemented in the .x.