Elm_Panel: forbid changing of content of "elm.swallow.event" part
Summary:
Because the "elm.swallow.event" part is used for internal needs of Panel widget
setting up or unsetting content of this part can break the behavior of this widget.
Due to the fact that this part is invisible and does not have any use cases for user
it seems acceptable to forbid changing of its content.

@fix

Reviewers: cedric, seoz, raster, Hermet

Reviewed By: Hermet

Subscribers: reutskiy.v.v

Differential Revision: https://phab.enlightenment.org/D1587