edje: Edje_Edit - fix segfault on copying state of IMAGE part with twins
Summary:
The copying of state of IMAGE part with twins caused the sefmentation fault due to the extra memory freing in _edje_edit_part_state_copy() that is deleted with this commit.
The data that is freed was previously copied from the structure of state 'from' and freeing of this data leads to freeing data of 'from' part.
Later, on copying of tweens to 'to' state array of tweens appears segmentation fault because we try to access freed data of 'from' state.

@fix

Reviewers: cedric, Hermet, raster, seoz

Subscribers: cedric, reutskiy.v.v

Differential Revision: https://phab.enlightenment.org/D1636

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>