Revert "Ecore File: Added test case for Ecore File Monitor"
This breaks make check and distcheck when run more than once. I first look
seems to indicate that the hardcoded /tmp/rootDir is not removed during
shutdown and fials to be created the next time as it is already present.

When removing it manually another problem with the subdir popped up. No
time to fix this just before I want to do the release so this gets
reverted. Please fix this up and submit it again for 1.13. Thanks.

This reverts commit cb84017a3f4ae6b3c13ee9fe7c0dca04a16aa43c.